\name{latticles}
\alias{latticles}
\title{Launch latticles GUI}
\usage{
latticles()
}
\description{
Importe artigos, classifique pelo Qualis-CAPES (2013-2016) e calcule metricas como Eq. A1 e quantitativo A1-A2-B1, tudo de forma automatica a partir de arquivos XML do curriculo Lattes.
}
\examples{
\dontrun{
library(shiny)
latticles()
}
}
\keyword{latticles}
